+++ app/graph/nodes/answer/struct_db_node.py
